requirements.txt
numpy~=1.20.3
scikit-image
pillow~=9.1.0
matplotlib~=3.5.1
sklearn~=0.0
scikit-learn~=1.0.2
tqdm~=4.61.0